Review request 10894 - added required signature in AccountProxy for playbook account preview usecase

+++ le-network/src/main/java/com/latticeengines/network/exposed/objectapi/AccountInterface.java
@@ -6,5 +6,5 @@ import com.latticeengines.domain.exposed.query.DataRequest;
 public interface AccountInterface {
     long getAccountsCount(String customerSpace, String start, DataRequest dataRequest);
 
-    DataPage getAccounts(String customerSpace, String start, Integer offset, Integer pageSize, DataRequest dataRequest);
+    DataPage getAccounts(String customerSpace, String start, Long offset, Long pageSize, DataRequest dataRequest);
 }
